+++ packages/google_maps_flutter/google_maps_flutter_web/lib/src/convert.dart
@@ -1021,9 +1021,20 @@ void _applyCameraUpdate(gmaps.Map map, CameraUpdate update) {
     case 'zoomTo':
       map.zoom = json[1]! as num;
     case 'newLatLngBoundsWithEdgeInsets':
-      throw UnsupportedError(
-        'CameraUpdate.newLatLngBoundsWithEdgeInsets is not supported on web. '
-        'Use CameraUpdate.newLatLngBounds with uniform padding instead.',
+      final List<Object?> latLngPair = asJsonList(json[1]);
+      final List<Object?> latLng1 = asJsonList(latLngPair[0]);
+      final List<Object?> latLng2 = asJsonList(latLngPair[1]);
+      map.fitBounds(
+        gmaps.LatLngBounds(
+          gmaps.LatLng(latLng1[0]! as num, latLng1[1]! as num),
+          gmaps.LatLng(latLng2[0]! as num, latLng2[1]! as num),
+        ),
+        gmaps.Padding(
+          top: json[2]! as num,
+          left: json[3]! as num,
+          bottom: json[4]! as num,
+          right: json[5]! as num,
+        ),
       );
     default:
       throw UnimplementedError('Unimplemented CameraMove: ${json[0]}.');